Services changes
- Moved the data to be loaded into src/main/resources/data to allow
classpath loading
- Added a BookingService, VenueService
- Services now have full functionality to create events, venues,
bookings and be able to process bookings with MRTs
- Added unit tests in.

+++ src/main/java/com/aerospike/demos/seat_reservation_demo/model/Venue.java
@@ -7,8 +7,10 @@ public class Venue {
     private String id;
     private String name;
     private String description;
-    private long numRows;
-    private long seatCount;
+    /** The number of rows of seats in this venue */
+    private int numRows;
+    /** The number of seats per row */
+    private int seatCount;
     private String address;
     private String city;
     private String country;
